colgon_plugin.py: rename colcon.py to colcon_plugin.py for consistency
All other snapcraft plugin files have a standard naming convention.
