+++ snapcraft/parts/plugins/__init__.py
@@ -17,7 +17,7 @@
 """Snapcraft specific plugins."""
 
 
-from .colcon import ColconPlugin
+from .colcon_plugin import ColconPlugin
 from .conda_plugin import CondaPlugin
 from .flutter_plugin import FlutterPlugin
 from .python_plugin import PythonPlugin
