+++ snapcraft/parts/plugins/register.py
@@ -18,7 +18,7 @@
 
 import craft_parts
 
-from .colcon import ColconPlugin
+from .colcon_plugin import ColconPlugin
 from .conda_plugin import CondaPlugin
 from .flutter_plugin import FlutterPlugin
 from .python_plugin import PythonPlugin
